+++ browser-tests/pages/harbors.ts
@@ -4,9 +4,9 @@ export const harbors = {
   harborList: {
     firstHarborLink: Selector('div[class^="harborListPage"] div[class^="table_rowWrapper"]:first-of-type a'),
   },
-  harborDetails: {
+  harborView: {
     address: Selector(
-      'div[class^="individualHarborPage"] div[class^="harborProperties_details"]:first-of-type span[class^="text_standard"]'
+      'div[class^="harborViewPage"] div[class^="harborProperties_details"]:first-of-type span[class^="text_standard"]'
     ),
   },
 };
