+++ src/features/harborView/harborViewPage/HarborViewPage.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import styles from './individualHarborPage.module.scss';
+import styles from './harborViewPage.module.scss';
 import PageTitle from '../../../common/pageTitle/PageTitle';
 
-const IndividualHarborPage: React.SFC = ({ children }) => {
+const HarborViewPage: React.SFC = ({ children }) => {
   const { t } = useTranslation();
   return (
-    <div className={styles.individualHarborPage}>
+    <div className={styles.harborViewPage}>
       <PageTitle title={t('individualHarbor.title')} />
       {children}
     </div>
   );
 };
 
-export default IndividualHarborPage;
+export default HarborViewPage;
