+++ src/features/harborView/harborViewPage/harborViewPage.module.scss
@@ -1,6 +1,6 @@
 @import 'spacings';
 
-.individualHarborPage {
+.harborViewPage {
   min-height: 100%;
   padding: units(2);
   display: flex;
